# server/docs/swagger.yaml
basePath: /
definitions:
...
  general.UserCollect:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      postsId:
        description: 帖子编号
        type: integer
      userId:
        description: 用户编号
        type: integer
    type: object
...
  request.IdDelete:
    properties:
      id:
        description: 编号
        type: integer
    type: object
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.UserCollectReq:
    properties:
      postsId:
        description: 帖子编号
        type: integer
    type: object
...
  response.PageResult:
    properties:
      list:
        type: object
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /app/userCollect/createUserCollect:
    post:
      consumes:
      - application/json
      parameters:
      - description: 收藏帖子
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserCollectReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 收藏帖子
      tags:
      - App_UserCollect
  /app/userCollect/deleteUserCollect:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除UserCollect
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdDelete'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除UserCollect
      tags:
      - App_UserCollect
  /app/userCollect/deleteUserCollectByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除UserCollect
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除UserCollect
      tags:
      - App_UserCollect
  /app/userCollect/getUserCollectList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回general.UserCollect
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/general.UserCollect'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取UserCollect列表
      tags:
      - App_UserCollect
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"